+++ crates/miden-testing/src/kernel_tests/tx/test_fpi.rs
@@ -19,7 +19,7 @@ use miden_objects::{
     FieldElement,
     account::{
         Account, AccountBuilder, AccountComponent, AccountProcedureInfo, AccountStorage,
-        PartialAccount, StorageSlot,
+        AccountStorageMode, PartialAccount, StorageSlot,
     },
     testing::{account_component::AccountMockComponent, storage::STORAGE_LEAVES_2},
     transaction::{AccountInputs, TransactionScript},
@@ -88,6 +88,7 @@ fn test_fpi_memory() {
             )
             .unwrap(),
         )
+        .storage_mode(AccountStorageMode::Public)
         .build_existing()
         .unwrap();
 
@@ -348,6 +349,7 @@ fn test_fpi_memory_two_accounts() {
             AccountMockComponent::new_with_empty_slots(TransactionKernel::testing_assembler())
                 .unwrap(),
         )
+        .storage_mode(AccountStorageMode::Public)
         .build_existing()
         .unwrap();
 
@@ -544,6 +546,7 @@ fn test_fpi_execute_foreign_procedure() {
             AccountMockComponent::new_with_slots(TransactionKernel::testing_assembler(), vec![])
                 .unwrap(),
         )
+        .storage_mode(AccountStorageMode::Public)
         .build_existing()
         .unwrap();
 
@@ -765,6 +768,7 @@ fn test_nested_fpi_cyclic_invocation() {
             AccountMockComponent::new_with_slots(TransactionKernel::testing_assembler(), vec![])
                 .unwrap(),
         )
+        .storage_mode(AccountStorageMode::Public)
         .build_existing()
         .unwrap();
 
@@ -956,6 +960,7 @@ fn test_nested_fpi_stack_overflow() {
                     )
                     .unwrap(),
                 )
+                .storage_mode(AccountStorageMode::Public)
                 .build_existing()
                 .unwrap();
 
@@ -1074,6 +1079,7 @@ fn test_nested_fpi_native_account_invocation() {
             AccountMockComponent::new_with_slots(TransactionKernel::testing_assembler(), vec![])
                 .unwrap(),
         )
+        .storage_mode(AccountStorageMode::Public)
         .build_existing()
         .unwrap();
 
@@ -1216,7 +1222,7 @@ fn test_fpi_stale_account() {
     // original unmodified foreign account. This should result in the foreign account's proof to be
     // invalid for this account tree root.
     let tx_context = mock_chain
-        .build_tx_context(native_account.id(), &[], &[])
+        .build_tx_context(native_account, &[], &[])
         .foreign_accounts(vec![overridden_foreign_account_inputs])
         .build();